backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11.1"
dependencies = [
    "music-tag>=0.4.3",
    "uvicorn>=0.34.0",
    "django>=4.2.20",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "black>=25.1.0",
]
